+++ src/django_components_js/src/index.ts
@@ -4,11 +4,16 @@ import { unescapeJs } from './utils';
 
 export type * from './manager';
 
-export const Components = {
+export const DjangoComponents = {
   manager: createComponentsManager(),
   createComponentsManager,
   unescapeJs,
 };
 
-// In browser, this is accessed as `Components.manager`, etc
-globalThis.Components = Components;
+// In browser, this is accessed as `DjangoComponents.manager`, etc
+globalThis.DjangoComponents = DjangoComponents;
+
+// TODO_v1: Delete this in v1, kept for backwards compatibility
+if (globalThis.Components === undefined) {
+  globalThis.Components = DjangoComponents;
+}